@@ -17,11 +17,11 @@
         "Memio\\Validator\\": "src/Memio/Validator"
     }},
     "require": {
-        "php": "^5.3|^7.0",
-        "memio/model": "^1.0"
+        "php": "^7.0",
+        "memio/model": "^2.0"
     },
     "require-dev": {
-        "fabpot/php-cs-fixer": "^1.6",
-        "phpspec/phpspec": "^2.4"
+        "friendsofphp/php-cs-fixer": "^2.10",
+        "phpspec/phpspec": "^4.3"
     }
 }
