@@ -19,25 +19,10 @@ use Memio\Validator\ViolationCollection;
 
 class ContractValidator implements ModelValidator
 {
-    /**
-     * @var CollectionValidator
-     */
     private $collectionValidator;
-
-    /**
-     * @var ConstraintValidator
-     */
     private $constraintValidator;
-
-    /**
-     * @var MethodValidator
-     */
     private $methodValidator;
 
-    /**
-     * @param CollectionValidator $collectionValidator
-     * @param MethodValidator     $methodValidator
-     */
     public function __construct(
         CollectionValidator $collectionValidator,
         MethodValidator $methodValidator
@@ -48,26 +33,17 @@ class ContractValidator implements ModelValidator
         $this->constraintValidator = new ConstraintValidator();
     }
 
-    /**
-     * {@inheritDoc}
-     */
     public function add(Constraint $constraint)
     {
         $this->constraintValidator->add($constraint);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($model)
+    public function supports($model): bool
     {
         return $model instanceof Contract;
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function validate($model)
+    public function validate($model): ViolationCollection
     {
         if (!$this->supports($model)) {
             return new ViolationCollection();
